Create main.py
Introducing SoftQR, one of my very first applications!

SoftQR is a simple Python-based app that allows users to locally generate QR codes. The app provides the ability to add custom logos or images to the QR codes, and it offers a convenient download option through an integrated web browser. Currently, it runs as an executable file.

I’m continuously working on improvements and plan to release updates with additional features in the near future.

Thank you for your support and interest!